Update db and fix bug

+++ ComputerProject/DETAIL_REPORT_REVENUE.cs
@@ -16,7 +16,7 @@ namespace ComputerProject
     {
         public int reportId { get; set; }
         public byte day { get; set; }
-        public int money { get; set; }
+        public long money { get; set; }
         public int amount { get; set; }
     
         public virtual REPORT REPORT { get; set; }
